Fixed: Book filters on author properties
Fixes #1755
